fix: ensure pullRequestCard still works in both stacking/non-stacking contexts
