+++ apps/desktop/src/lib/pr/PullRequestCard.svelte
@@ -143,8 +143,10 @@
 				tooltip="Open in browser"
 				onclick={() => {
 					openExternalUrl($pr.htmlUrl);
-				}}>View PR</Button
+				}}
 			>
+				View PR
+			</Button>
 		</div>
 
 		<!--
